fix: input types for form
